requirements.txt
fastapi
uvicorn
torch
torchvision
python-multipart
pillow
torch-geometric
scikit-learn